+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add the version number to html of the root div as a data attribute
 - Add a changelog
 - Single source of truth for the version and improved footer with questions' version
+- Change scores ranks and add a section for ppl that look for the maximum
 
 
 ## [0.1.0] - 2020-12-18